pyproject.toml
[project]
name = "robots"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "lerobot[lekiwi]",
    "feetech-servo-sdk",
    "python-dotenv",
    "wandb",
    "accelerate>=1.12.0",
    "pygame>=2.6.1",
    "transformers>=4.57.6",
    "num2words>=0.5.14",
]

[project.optional-dependencies]
vla = [
    "lerobot[smolvla]",
]
